Fix many header problems and warnings (new clang/libstdc++). Add a few useful things to Util.
git-svn-id: file:///home/runner/work/tom7misc-sync-action/tom7misc-sync-action/upstream/trunk@6620 c7e11af0-4c7a-11f1-9edd-117cb6f09d55

+++ cc-lib/ansi.cc
@@ -270,7 +270,7 @@ std::string ANSI::ProgressBar(uint64_t numer, uint64_t denom,
 
   // The length when we composite is governed by the fg/bg raster, so
   // make sure these are the right length.
-  if (fgraster.size() > bar_width) {
+  if ((int)fgraster.size() > bar_width) {
     fgraster.resize(bar_width);
   } else {
     while ((int)fgraster.size() < bar_width) {
